+++ src/test/java/roomescape/reservationwaiting/domain/ReservationWaitingTest.java
@@ -9,7 +9,7 @@ import java.time.ZoneId;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.common.domain.ReservationSlot;
-import roomescape.exception.BusinessException;
+import roomescape.common.exception.BusinessException;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
